+++ App/Pages/Hub.xaml.cs
@@ -1,0 +1,17 @@
+using Microsoft.Extensions.Logging;
+
+
+namespace TMS_APP.Pages
+{
+    public partial class Hub : ContentPage
+    {
+        private readonly ILogger<Hub> _logger;
+        public Hub(ILogger<Hub> logger)
+        {
+            InitializeComponent();
+            _logger = logger ?? throw new ArgumentNullException(nameof(logger));
+        }
+    }
+}
+
+
