+++ App/Platforms/Android/AndroidManifest.xml
