+++ App/Platforms/Android/MainActivity.cs
