+++ App/Platforms/Android/MainApplication.cs
