+++ App/Platforms/Android/Resources/values/colors.xml
