+++ App/Platforms/MacCatalyst/AppDelegate.cs
