+++ App/Platforms/MacCatalyst/Entitlements.plist
