+++ App/Platforms/MacCatalyst/Info.plist
