+++ App/Platforms/Tizen/Main.cs
