+++ App/Platforms/Tizen/tizen-manifest.xml
