+++ App/Platforms/Windows/App.xaml.cs
