+++ App/Resources/Images/Icons/stops/stopIcon-large.png
