+++ App/Resources/Images/Icons/stops/stopIcon_small.png
