+++ App/Resources/Images/dotnet_bot.png
