+++ App/Resources/Raw/AboutAssets.txt
