+++ TMS_APP.sln
@@ -3,7 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.5.002.0
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TMS_APP", "TMS_APP.csproj", "{BE9186FD-6C02-44CC-B336-C19562218377}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TMS_APP", "App\TMS_APP.csproj", "{BE9186FD-6C02-44CC-B336-C19562218377}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Tests", "Tests\Tests.csproj", "{BD2AB876-5629-4A0C-BEE7-86E1F8D98EFD}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +17,10 @@ Global
 		{BE9186FD-6C02-44CC-B336-C19562218377}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{BE9186FD-6C02-44CC-B336-C19562218377}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{BE9186FD-6C02-44CC-B336-C19562218377}.Release|Any CPU.Build.0 = Release|Any CPU
+		{BD2AB876-5629-4A0C-BEE7-86E1F8D98EFD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{BD2AB876-5629-4A0C-BEE7-86E1F8D98EFD}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{BD2AB876-5629-4A0C-BEE7-86E1F8D98EFD}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{BD2AB876-5629-4A0C-BEE7-86E1F8D98EFD}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
